console: Add wrappers for many protocol functions
Provide associated functions as alternatives to directly calling the
C-based functions in the protocol definitions. This improves the
ergonomics of writing Rust code by allowing the use of `Result` and the
try shorthand (`?`) with functions.

This includes a fix to the function signatures of RegisterKeyNotify and
UnregisterKeyNotify. The final argument is called "Handle" but is a void
pointer to an unspecified structure and not an EFI_HANDLE.

Signed-off-by: Tim Crawford <tcrawford@system76.com>